requirements.txt
JPype1==1.6.0
numpy==1.24.3
pandas==2.0.3
scipy==1.16.3
